requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
xgboost==1.7.6
lightgbm==4.1.0
plotly==5.17.0
yfinance==0.2.28
python-dotenv==1.0.0
google-generativeai==0.8.3